BaladosSyncProjections.Projectors.SubscriptionsProjector (Balados Sync v0.1.0)

Functions

child_spec(opts)

Provides a child specification to allow the event handler to be easily supervised.

Supports the same options as start_link/3.

The default options supported by GenServer.start_link/3 are also supported, including the :hibernate_after option which allows the process to go into hibernation after a period of inactivity.

Example

Supervisor.start_link([
  {ExampleHandler, []}
], strategy: :one_for_one)

schema_prefix(event)

Callback implementation for Commanded.Projections.Ecto.schema_prefix/1.

schema_prefix(event, metadata)

Callback implementation for Commanded.Projections.Ecto.schema_prefix/2.

start_link(opts \\ [])

Start an event handler GenServer process linked to the current process.

Options

:application - the Commanded application.

:name - name of the event handler used to determine its unique event store subscription.

:concurrency - determines how many processes are started to concurrently process events. The default is one process.

:consistency - one of either :eventual (default) or :strong.

:start_from - where to start the event store subscription from when first created (default: :origin).

:subscribe_to - which stream to subscribe to can be either :all to subscribe to all events or a named stream (default: :all).

:batch_size - the size of batches to deliver to handle_batch in batched mode.

The default options supported by GenServer.start_link/3 are supported, including the :hibernate_after option which allows the process to go into hibernation after a period of inactivity.

update_projection(event, metadata, multi_fn)